requirements.txt
oracledb
sentence-transformers
pandas
python-dotenv